#7 Move request, response from interfaces to api
