+++ apps/client/api/index.ts
@@ -1,5 +1,16 @@
-import axios from 'axios'
-import { RequestProps } from 'interfaces'
+import axios, { AxiosRequestConfig, Method, AxiosResponse, AxiosError } from 'axios'
+
+export type ResponseSuccessHandler<T = unknown> = (res: AxiosResponse<T>) => Promise<void> | void
+export type ResponseErrorHandler<T = unknown> = (err: AxiosError<T>) => Promise<void> | void
+
+export interface RequestProps<T = unknown> {
+  method?: Method
+  url: string
+  data?: unknown
+  config?: AxiosRequestConfig
+  onSuccess: ResponseSuccessHandler<T>
+  onError: ResponseErrorHandler<T>
+}
 
 const customAxios = axios.create({
   baseURL: `/api`,
