+++ apps/client/interfaces/index.ts
@@ -1,13 +1,4 @@
-import { RequestProps } from './request'
-import { ResponseSuccessHandler, ResponseErrorHandler } from './response'
 import { Token } from './token'
 import { KakaoProfile, User } from './user'
 
-export type {
-  ResponseSuccessHandler,
-  ResponseErrorHandler,
-  RequestProps,
-  Token,
-  KakaoProfile,
-  User,
-}
+export type { Token, KakaoProfile, User }
